refactor/ProductDetail 컴포넌트 1차 정리
